api/requirements.txt
fastapi
uvicorn
requests
PyJWT
redis
mlflow
numpy
yfinance
tensorflow
boto3
python-multipart 
python-jose[cryptography]